Step 12: Change PlatformWord from usize to cfg-based u64/u32
Remove the compile_error! gate for non-64-bit targets. PlatformWord is
now u64 on 64-bit and u32 on 32-bit. All inventory/spill arrays,
SmallCounters, and SelectSmall use PlatformWord with explicit casts at
read/write boundaries. Rank9/Select9 remain pinned to u64.

Dense PartialArray now uses Box<[u64]> backing (Rank9 requires u64).
SelectSmall/SelectZeroSmall SUPERBLOCK_BIT_SIZE uses cfg to avoid
1usize << 32 overflow on 32-bit. thread-priority made optional behind
the rayon feature for wasm portability.

Cross-compiles successfully for wasm32-wasip1.

+++ src/bits/bit_vec.rs
@@ -690,8 +690,7 @@ macro_rules! impl_rank_hinted_for_bitvec {
     };
 }
 
-// TODO(32-bit): when PlatformWord changes to u64, change usize to u64 here
-impl_rank_hinted_for_bitvec!(usize, 64);
+impl_rank_hinted_for_bitvec!(u64, 64);
 impl_rank_hinted_for_bitvec!(u32, 32);
 
 // SelectHinted and SelectZeroHinted for PlatformWord-backed bit vectors.
